Pull our create Range Position function
